pyproject.toml
[project]
name = "personal-finance-tracker"
version = "0.1.0"
description = "Personal finance tracker using Plaid API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.0",
    "plaid-python>=34.0.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "ruff>=0.12.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]